third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/post-blob-content-type-sync-expected.txt

CONSOLE WARNING: line 49: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
CONSOLE ERROR: line 52: Failed to load http://localhost:8000/xmlhttprequest/resources/access-control-allow-lists.php: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
Test verifies that content MIME type is set correctly when Blob is sent using XMLHttpRequest synchronously.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS postedMimeType is "text/plain;charset=utf-8"
PASS postedMimeType === undefined is true
PASS postedMimeType === undefined is true
PASS postedMimeType === undefined is true
PASS postedMimeType === undefined is true
PASS postedMimeType === undefined is true
PASS postedMimeType === undefined is true
PASS postedMimeType === undefined is true
PASS postedMimeType is "multipart/mixed;boundary=\"--blob-boundary\""
PASS postedMimeType === undefined is true
PASS postedMimeType === undefined is true
PASS postedMimeType === undefined is true
PASS Exception should be thrown.
PASS postedMimeType is "text/plain;charset=utf-8"
PASS postedMimeType is "text/plain;charset=utf-8"
PASS postedMimeType === undefined is true
PASS postedMimeType === undefined is true
PASS postedMimeType is "multipart/mixed;boundary=\"--blob-boundary\""
PASS postedMimeType === undefined is true
PASS postedMimeType === undefined is true
PASS successfullyParsed is true

TEST COMPLETE
